+++ middleware/multicore/rpmsg_lite/lib/include/virtqueue.h
@@ -95,7 +95,7 @@ struct virtqueue
 {
     /* 32bit aligned { */
     char vq_name[VIRTQUEUE_MAX_NAME_SZ];
-    uint32_t vq_flags;
+    unsigned long vq_flags;
     int vq_alignment;
     int vq_ring_size;
     void *vq_ring_mem;
@@ -107,29 +107,29 @@ struct virtqueue
     /* } 32bit aligned */
 
     /* 16bit aligned { */
-    uint16_t vq_queue_index;
-    uint16_t vq_nentries;
-    uint16_t vq_free_cnt;
-    uint16_t vq_queued_cnt;
+    unsigned short vq_queue_index;
+    unsigned short vq_nentries;
+    unsigned short vq_free_cnt;
+    unsigned short vq_queued_cnt;
 
     /*
      * Head of the free chain in the descriptor table. If
      * there are no free descriptors, this will be set to
      * VQ_RING_DESC_CHAIN_END.
      */
-    uint16_t vq_desc_head_idx;
+    unsigned short vq_desc_head_idx;
 
     /*
      * Last consumed descriptor in the used table,
      * trails vq_ring.used->idx.
      */
-    uint16_t vq_used_cons_idx;
+    unsigned short vq_used_cons_idx;
 
     /*
      * Last consumed descriptor in the available table -
      * used by the consumer side.
      */
-    uint16_t vq_available_idx;
+    unsigned short vq_available_idx;
     /* } 16bit aligned */
 
     boolean avail_read;  /* 8bit wide */
@@ -137,7 +137,7 @@ struct virtqueue
     boolean used_read;   /* 8bit wide */
     boolean used_write;  /* 8bit wide */
 
-    uint16_t padd; /* aligned to 32bits after this: */
+    unsigned short padd; /* aligned to 32bits after this: */
 
     void *priv; /* private pointer, upper layer instance pointer */
 #if defined(RL_USE_ENVIRONMENT_CONTEXT) && (RL_USE_ENVIRONMENT_CONTEXT == 1)
@@ -149,9 +149,9 @@ struct virtqueue
 struct vring_alloc_info
 {
     void *phy_addr;
-    uint32_t align;
-    uint16_t num_descs;
-    uint16_t pad;
+    unsigned long align;
+    unsigned short num_descs;
+    unsigned short pad;
 };
 
 struct vq_static_context
@@ -217,17 +217,17 @@ int virtqueue_create_static(unsigned short id,
                             struct virtqueue **v_queue,
                             struct vq_static_context *vq_ctxt);
 
-int virtqueue_add_buffer(struct virtqueue *vq, uint16_t head_idx);
+int virtqueue_add_buffer(struct virtqueue *vq, unsigned short head_idx);
 
-int virtqueue_fill_used_buffers(struct virtqueue *vq, void *buffer, uint32_t len);
+int virtqueue_fill_used_buffers(struct virtqueue *vq, void *buffer, unsigned long len);
 
-int virtqueue_fill_avail_buffers(struct virtqueue *vq, void *buffer, uint32_t len);
+int virtqueue_fill_avail_buffers(struct virtqueue *vq, void *buffer, unsigned long len);
 
-void *virtqueue_get_buffer(struct virtqueue *vq, uint32_t *len, uint16_t *idx);
+void *virtqueue_get_buffer(struct virtqueue *vq, unsigned long *len, unsigned short *idx);
 
-void *virtqueue_get_available_buffer(struct virtqueue *vq, uint16_t *avail_idx, uint32_t *len);
+void *virtqueue_get_available_buffer(struct virtqueue *vq, unsigned short *avail_idx, unsigned long *len);
 
-int virtqueue_add_consumed_buffer(struct virtqueue *vq, uint16_t head_idx, uint32_t len);
+int virtqueue_add_consumed_buffer(struct virtqueue *vq, unsigned short head_idx, unsigned long len);
 
 void virtqueue_disable_cb(struct virtqueue *vq);
 
@@ -243,9 +243,9 @@ void virtqueue_dump(struct virtqueue *vq);
 
 void virtqueue_notification(struct virtqueue *vq);
 
-uint32_t virtqueue_get_desc_size(struct virtqueue *vq);
+unsigned long virtqueue_get_desc_size(struct virtqueue *vq);
 
-uint32_t virtqueue_get_buffer_length(struct virtqueue *vq, uint16_t idx);
+unsigned long virtqueue_get_buffer_length(struct virtqueue *vq, unsigned short idx);
 
 void vq_ring_init(struct virtqueue *vq);
